motor scripting language is coming along

+++ firmware/sine.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int16_t cosine(int16_t t, int16_t amp);
-int16_t sine(int16_t t, int16_t amp);
+int16_t cosine(float t, float amp);
+int16_t sine(float t, float amp);
 
 #endif
